Apple의 타입을 double로 변경

+++ src/practice/Apple.java
@@ -1,9 +1,9 @@
 package practice;
 
 public class Apple {
-	int a;
-	int b;
-	int c;
+	double a;
+	double b;
+	double c;
 	Float d;
 	
 	public Apple(int a, int b, int c) {
